requirements.txt
billiard==3.6.4.0
black==22.3.0
coreschema==0.0.4
coverage==6.4.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.13.0
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-yasg==1.21.3
flake8==5.0.2
geographiclib==1.52
geopy==2.2.0
isort==5.10.1
nose==1.3.7
openapi-codec==1.3.2
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
redis==4.3.3


